Вынесли объект date в отдельный файл. Исправили пути
