+++ src/consts.js
@@ -13,4 +13,7 @@ const NEW_EVENT = {
   type: 'flight',
 };
 
-export {EVENT_TYPES, PLACES, EVENTS_NUMBER, NEW_EVENT};
+const HOURS_IN_DAY = 24;
+const MINUTES_IN_HOUR = 60;
+
+export {EVENT_TYPES, PLACES, EVENTS_NUMBER, NEW_EVENT, HOURS_IN_DAY, MINUTES_IN_HOUR};
